Add skeleton code for the spellchecker
This is to define the API so I can write tests against it. Note that I
chose to pass an explicit alphabet to inserts, replaces, known_edits,
and known. This has two goals. The first is that it is a more of a
functional style, as is fitting for modern javascript. The second is
that it's easier to test.